add new series and move it to data folder
