+++ server/model/article.js
@@ -5,7 +5,8 @@ import {type Person} from './person';
 import {type Picture} from './picture';
 import {type ContentType} from './content-type';
 import {type Video} from './video';
-import {type ArticleSeries, getSeriesCommissionedLength} from './series';
+import {type ArticleSeries} from './series';
+import {getSeriesCommissionedLength} from '../data/series';
 import {getWpFeaturedImage} from './media';
 import {bodyParser} from '../util/body-parser';
 import {authorMap} from '../services/author-lookup';
